# src/docs/swagger.yaml
openapi: 3.0.0
info:
    title: API Documentation
    version: 1.0.0
    description: This is a sample API documentation using OpenAPI Specification.
paths:
    /api/users:
        get:
            summary: Get all users
            responses:
                "200":
                    description: A list of users
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        id:
                                            type: integer
                                        name:
                                            type: string
                                        email:
                                            type: string
    /api/users/{id}:
        get:
            summary: Get a user by ID
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                  description: The ID of the user to retrieve
            responses:
                "200":
                    description: A user object
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                    name:
                                        type: string
                                    email:
                                        type: string
                "404":
                    description: User not found
    /api/users/{id}/documents:
        post:
            summary: Create a document for a user
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                  description: The ID of the user to create a document for
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                content:
                                    type: string
            responses:
                "201":
                    description: Document created successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: integer
                                    title:
                                        type: string
                                    content:
                                        type: string
                "400":
                    description: Invalid input
    /api/pets:
        get:
            summary: Get all pets
            responses:
                "200":
                    description: A list of pets
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        id:
                                            type: integer
                                        name:
                                            type: string
                                        type:
                                            type: string